pallet_authors::pallet::dispatchables

Function create_pool

pub fn create_pool<T: Config>(index: IndexDigest<T>, commission: Commission<T>)

Create a new funding pool backed by an existing index.

Pools enable managed aggregation of funds with an explicit commission applied to rewards earned via the underlying index.

Pool configuration is uniquely identified by its digest.

Commission is fixed at creation time, changing commission after creates a new pool with same slots and updated commission.

The caller becomes the initial pool manager.

Emits: Event::PoolCreated via deposit event

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::create_pool.