pallet_authors::pallet::dispatchables

Function enlist

pub fn enlist<T: Config>(
    collateral: AuthorAsset<T>,
    fortitude: FortitudeWrapper,
)

Enlist the caller as an author by locking the required collateral.

Establishes the caller as an author by placing collateral under the author role commitment.

Requirements

Provided collateral must be at least MinCollateral.

Fails if the caller is already enrolled as an author.

Behavior

Locks the required collateral and associates it with the author role.

Ensures at least the minimum collateral is maintained.

Execution Controls

fortitude defines how the collateral is sourced:

FortitudeWrapper::Force: Uses the caller’s liquid balance to place the collateral.

FortitudeWrapper::Polite: Uses funds already deposited into the commitment reserve (if Config::CommitmentAdapter == pallet_commitment).

Emits: Event::AuthorEnlisted

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::enlist.