pallet_authors::pallet::dispatchables

Function refill

pub fn refill<T: Config>(
    collateral: AuthorAsset<T>,
    fortitude: FortitudeWrapper,
)

Increase the caller’s collateral by locking additional assets.

Adds to the existing collateral, strengthening the author’s position and ensuring compliance with evolving system requirements such as MinCollateral.

Behavior

Collateral is added on top of existing locked collateral.

Execution Controls

fortitude defines how the collateral is sourced:

FortitudeWrapper::Force: Uses the caller’s liquid balance to place the additional collateral.

FortitudeWrapper::Polite: Uses funds already deposited into the commitment reserve (if Config::CommitmentAdapter == pallet_commitment).

Emits: Event::AuthorCollateralRaised

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::refill.