pallet_authors::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AuthorFunders

Stores the funder details for each (Author, Backer) pair.

AuthorPenalties

Stores pending penalties for authors by ([BlockNumberFor], Author) key.

AuthorRewards

Stores pending rewards for authors by ([BlockNumberFor], Author) key.

AuthorsDigest

Maps each AuthorDigest to its corresponding Author account.

AuthorsMap

Maps each Author account to its on-chain metadata (AuthorInfo).

Elected

Persistent record of elected authors per block, representing the outcome of each finalized election round.

ForceMaxElected

A strictness flag controlling how the system enforces the MaxElected limit.

IncreaseProbationBy

Number of blocks by which an author’s risk period is increased upon misbehavior.

MaxElected

The upper bound on the number of authors that can be elected in a single election round.

MaxExposure

Global Maximum exposure allowed globally per funding operation.

MinCollateral

Global Minimum collateral required for any author.

MinElected

The minimum required number of authors that must be elected for an election to be considered valid.

MinFund

Global Minimum funding required globally per funding operation.

PenaltiesBuffer

Buffer period in blocks before a penalty can be applied.

PenaltiesUntil

Tracks the latest block number until which author penalties are scheduled.

ProbationPeriod

Number of blocks representing the probation period for newly enrolled or demoted authors.

RecentElectedOn

Tracks the most recent block number in which an election was conducted and successfully stored in Elected.

ReduceProbationBy

Number of blocks by which an author’s risk period is reduced upon good behavior.

RewardsBuffer

Buffer period in blocks before a reward can be applied.

RewardsUntil

Tracks the latest block number until which author rewards are scheduled.