pallet_authors::pallet

Type Alias AuthorsMap

pub type AuthorsMap<T: Config> = StorageMap<_GeneratedPrefixForStorageAuthorsMap<T>, Blake2_128Concat, Author<T>, AuthorInfo<T>, OptionQuery>;

Maps each Author account to its on-chain metadata (AuthorInfo).

This storage is the primary record for all authors, tracking their:

lifecycle status (Active, Probation, Resigned),

risk period and timestamps,

funding constraints (min_fund, max_fund).

Keys are insert-mutate-only and MUST NOT be removed.

Indexes and pools may retain digests indefinitely, and re-enrollment of resigned authors may utilize their previous commitment-digest (existing meta).

By keeping the resigned authors meta-data alive funders can draw from resigned authors.

In Future, if casual resignations are penalized via assigning new commitment- digests, then this storage can be mutated again without reuse of digest, but removable clearly requires withdrawal, and usage in indexes and pools must not create inconsistent state.

Used for all author-related operations.

Storage type is [StorageMap] with key type Author < T > and value type AuthorInfo < T >.

Aliased Type

pub struct AuthorsMap<T: Config>(/* private fields */);