pallet_authors

Module types

Core types and aliases for the Authors system.

This module defines the primary structures and type aliases used by pallet_authors. These types are publicly exposed and used across the pallet’s APIs for representing Author-related data.

Trait implementations provided by this crate’s Pallet can use these types via trait-bound equality constraints to ensure type alignment with this pallet’s concrete implementations if necessary.

Example

mod pallet {
    use pallet_authors::types::AuthorInfo;

    pub trait Config: frame_system::Config {
        type RoleAdapter: RoleManager<Meta = AuthorInfo<Self>>;
    }
}

Structs

AuthorInfo

Represents the on-chain record of an Author role within the runtime.

Enums

AuthorStatus

Represents the lifecycle state of an Author role within the runtime.

ForceGenesisConfig

Enumerates configurable runtime-stored parameters that influences probation, rewards, penalties, election constraints, or economic limits may be forcibly overridden at runtime through privileged (root/governance) operations.

FortitudeWrapper

Wrapper type for [Fortitude] used in extrinsics.

Funder

Represents the entity responsible for backing or funding an Author role.

FundingTarget

Specifies the destination or mechanism by which funds are allocated within the authors / roles system.

PrecisionWrapper

Wrapper type for [Precision] used in extrinsics.

Type Aliases

Author

Represents an on-chain account that can hold an Author role.

AuthorAsset

Represents the asset or balance type associated with an Author in the context of funding, staking, or collateral.

AuthorDigest

Represents a cryptographic digest, a unique identifier for an Author.

Backer

Represents the account that provides backing to an Author, either directly or through an index or pool funding mechanism.

BackingElectionWeight

A single backing contribution used in backing-based election.

Commission

Commission Type used by the Pallet for Pool Commissions.

ElectViaBacking

Election input for backing-based selection.

ElectViaInfluence

Election input for influence-based selection.

ElectedAuthors

A list of authors selected by an election round.

IndexDigest

Represents a digest identifier for an indexed commitment (unmanaged pools) backing an Author role.

PoolDigest

Represents a digest identifier for a pool commitment (managed collective commitment) funding source for an Author role.

Ratio

Penalty ratio type used by the Author Roles subsystem.

Shares

Shares Type used by the Pallet for Index/Pool shares.